otl-prom-fastapi/requirements.txt
fastapi
uvicorn
opentelemetry-sdk
opentelemetry-api
opentelemetry-exporter-prometheus